began to implement optimal path finding
optimal path finding started, didn't seem worth the effort, currently not sure if maze is being closed correctly post first run as the mouse escapes and moves in odd directions after. I'll leave this as a branch if we have free time to figure it out but implementation of robot maze solving class is much more important at the moment. This will be the only place where optimal path finding is saved as I'll be rolling back on my local files.